AuthService 추가, GrantedRole 변경, JwtUtils 기능 추가, Diary 엔티티 변경
회원 가입 및 토큰 발급을 위한 AuthService 추가
회원 권한과 관련된 GrantedRole Enum명 변경
JwtUtils에 RefreshToken 발급을 위한 기능 추가
Diary 엔티티 필드 길이 변경

+++ src/test/java/com/farmdiary/api/security/jwt/AuthTokenFilterTest.java
@@ -57,7 +57,7 @@ class AuthTokenFilterTest {
 
     private final String email = "email@email.com";
     private final String jwtSecret = "secret";
-    private final int jwtExpirationMs = 864000;
+    private final long jwtExpirationMs = 864000;
 
     @BeforeEach
     void setUp() {
